ai/requirements.txt
# Core Framework
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0

# AI & LLM
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
httpx>=0.27.0